apt-requirements-bionic.txt
clang-8
clang-format-8
clang-tidy-8
cmake
git
libargue-dev
libeigen3-dev
libelf-dev
libgoogle-glog-dev
libgtk-3-dev
libgtkmm-3.0-dev
libprotobuf-dev
libtangent-json-dev
libtangent-util-dev
libtinyxml2-dev
libudev-dev
ninja-build
pkg-config
protobuf-compiler
python-pip
python3-minimal
python3-pip
shellcheck
